Move components into geoserve output library
