+++ projects/geoserve-output/src/lib/authoritative-region/authoritative-region.component.ts
