+++ projects/geoserve-output/src/lib/geoserve-output.component.ts
@@ -1,19 +1,29 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material';
+
+import {
+  CoordinatesService,
+  LocationDialogComponent
+} from 'location-input';
 
 @Component({
-  selector: 'geoserve-output',
-  template: `
-    <p>
-      geoserve-output works!
-    </p>
-  `,
-  styles: []
+  selector: 'app-geoserve-output',
+  templateUrl: './geoserve-output.component.html',
+  styleUrls: ['./geoserve-output.component.css']
 })
 export class GeoserveOutputComponent implements OnInit {
 
-  constructor() { }
+  constructor (
+    public coordinatesService: CoordinatesService,
+    public dialog: MatDialog
+  ) { }
 
   ngOnInit() {
   }
 
+  onClick(): void {
+    if (this.dialog && LocationDialogComponent) {
+      this.dialog.open(LocationDialogComponent);
+    }
+  }
 }
