+++ projects/geoserve-output/src/lib/location-output/location-output.component.ts
