+++ projects/geoserve-output/src/lib/nearby-place/nearby-place.component.css
