+++ projects/geoserve-output/src/lib/no-data/no-data.component.ts
