+++ projects/geoserve-output/src/lib/offshore-region/offshore-region.component.ts
