+++ projects/geoserve-output/src/lib/tectonic-summary-region/tectonic-summary-region.component.ts
