+++ projects/geoserve-output/src/public_api.ts
@@ -2,6 +2,5 @@
  * Public API Surface of geoserve-output
  */
 
-export * from './lib/geoserve-output.service';
 export * from './lib/geoserve-output.component';
 export * from './lib/geoserve-output.module';
